pipeline generation as plugins
