+++ depthai_ros_driver/CMakeLists.txt
@@ -52,6 +52,7 @@ depthai
 rclcpp 
 rclcpp_components 
 std_srvs
+pluginlib
 )
 
 
@@ -103,7 +104,8 @@ target_link_libraries(
 add_library(
   ${PROJECT_NAME} SHARED
   src/camera.cpp
-  src/pipeline_generator.cpp
+  src/pipeline/pipeline_generator.cpp
+  src/pipeline/base_types.cpp
 )
 
 ament_target_dependencies(${PROJECT_NAME} ${CAM_DEPS})
@@ -116,17 +118,27 @@ target_link_libraries(
 
 
 rclcpp_components_register_nodes(${PROJECT_NAME} "${PROJECT_NAME}::Camera")
+pluginlib_export_plugin_description_file(${PROJECT_NAME} plugins.xml)
 ament_export_include_directories(include)
+
+ament_export_targets(${PROJECT_NAME}Targets HAS_LIBRARY_TARGET)
+ament_export_libraries(${PROJECT_NAME})
+ament_export_dependencies(${COMMON_DEPS} ${SENSOR_DEPS} ${NN_DEPS} ${CAM_DEPS})
 
 install(DIRECTORY launch config DESTINATION share/${PROJECT_NAME})
 
-install(
-  TARGETS
-  ${PROJECT_NAME} ${COMMON_LIB_NAME} ${SENSOR_LIB_NAME} ${NN_LIB_NAME}
-  ARCHIVE DESTINATION lib
-  LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin
+install(TARGETS ${PROJECT_NAME}  ${SENSOR_LIB_NAME}
+${NN_LIB_NAME}
+${COMMON_LIB_NAME}
+EXPORT ${PROJECT_NAME}Targets
+        ARCHIVE DESTINATION lib
+        LIBRARY DESTINATION lib
+        RUNTIME DESTINATION bin
+        INCLUDES DESTINATION include
 )
+
+install(EXPORT ${PROJECT_NAME}Targets
+      DESTINATION share/${PROJECT_NAME}/cmake)
 
 ament_python_install_package(${PROJECT_NAME})
 # Install Python executables
@@ -139,6 +151,7 @@ install(
 ament_export_include_directories(
   include
 )
+
 install(
     DIRECTORY include/
     DESTINATION include
