+++ depthai_ros_driver/package.xml
@@ -28,6 +28,7 @@
   <depend>depthai_descriptions</depend>
   <depend>depthai_ros_msgs</depend>
   <depend>depthai_examples</depend>
+  <depend>pluginlib</depend>
   <export>
     <build_type>ament_cmake</build_type>
   </export>
